feat(notification): 알림 REST API 및 WebSocket 인앱 알림 구현

+++ notification-service/src/main/java/com/apten/notification/application/model/response/NotificationUnreadCountRes.java
@@ -11,5 +11,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class NotificationUnreadCountRes {
-    private Integer unreadCount;
+    // 현재 로그인 사용자의 미읽음 알림 개수
+    private Long unreadCount;
 }
